Depl-218574 Synchronize monitoring config with trackor.onevizion.com
